Moved the HashFunction to Util.
