+++ include/identity-based/PublicParameters.h
@@ -5,7 +5,7 @@
 
 #include "elliptic/AffinePoint.h"
 #include "elliptic/EllipticCurve.h"
-#include "identity-based/HashFunction.h"
+#include "util/HashFunction.h"
 
 
 // References
